Remove migrations initializer
